requirements.txt
numpy
pillow
matplotlib
tensorflow==1.8.0